Ran tests to verify the implemented authentication and authorization via local/jwt strategies for each endpoint. For now, the backend is at a functional point! Can go ahead and begin work on the Front-End admin view now.

+++ server/controllers/commentController.js
@@ -28,13 +28,15 @@ const createComment = asyncHandler(async (req, res) => {
             error: errors.array()
         })
     }
-    const {body, authorId, postId} = req.body
+
+    const client = req.user
+    const {body, postId} = req.body
     try {    
         const newComment = await prisma.comments.create(
             {
                 data: {
                     body,
-                    author: {connect: {id: authorId}},
+                    author: {connect: {id: client.id}},
                     post: {connect: {id: postId}}
                 }
             }
@@ -55,14 +57,30 @@ const editComment = asyncHandler(async (req, res) => {
             error: errors.array()
         })
     }
-
+    const client = req.user
     const {id, body}  = req.body
+
+    //Check the comment we're about to update to see if it's authorId matches the client.id
+    const commentToBeUpdated = await prisma.comments.findUnique(
+        {
+            where: {id: id}
+        }
+    )
+
+    if (!(commentToBeUpdated.authorId === client.id)) {
+        return res.status(403).json(
+            {
+                error: "Forbidden: Can't update another user's comment"
+            }
+        )
+    }
+    
     try {
         const comment = await prisma.comments.update({
             where: {
                 id: id
             },
-            data: body
+            data: {body: body}
         })
         res.status(200).json(comment)
     } catch (error) {
@@ -72,12 +90,30 @@ const editComment = asyncHandler(async (req, res) => {
 
 const deleteComment = asyncHandler(async (req, res) => {
     const id = req.body.id
+    const client = req.user
+
+    //grab the comment that we're about to delete
+    const commentToBeDeleted = await prisma.comments.findUnique(
+        {
+            where: {id: id}
+        }
+    )
+
+    //Check if the person making the request is the same as the author od the comment
+    if(!(commentToBeDeleted.authorId === client.id)) {
+        return res.status(403).json({
+            error: "Forbidden: Can't delete another user's comment"
+        })
+    }
+
+    //Otherwise, carry on
     try {
         await prisma.comments.delete({
             where: {
                 id: id
             }
         })
+        res.status(200).json({body: "Comment Deleted"})
     } catch (error) {
         res.status(400).json({error: `failed to delete comment: ${error}`})
     }
